Feat/order

+++ src/main/java/com/demo/myshop/model/Cart.java
@@ -23,7 +23,7 @@ public class Cart {
     @JoinColumn(name = "user_id")
     private User user;
 
-    private Double totalPrice = 0.0; // 총 가격
+    private Integer totalPrice = 0; // 총 가격
 
     @OneToMany(mappedBy = "cart", cascade = CascadeType.ALL, orphanRemoval = true)
     private Set<CartItem> items = new HashSet<>();
@@ -31,7 +31,7 @@ public class Cart {
     // 카트의 총 가격을 업데이트하는 메서드
     public void updateTotalPrice() {
         totalPrice = items.stream()
-                .mapToDouble(item -> item.getProduct().getPrice() * item.getQuantity())
+                .mapToInt(item -> item.getProduct().getPrice() * item.getQuantity())
                 .sum();
     }
 }
